backend/requirements.txt
annotated-types==0.7.0
asgiref==3.11.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
contextlib2==21.6.0
cryptography==46.0.3
Django==6.0
django-ninja==1.5.1
django-ninja-extra==0.30.8
django-ninja-jwt==5.4.2
fire==0.7.1
idna==3.11
injector==0.23.0
markdown-it-py==4.0.0
mdurl==0.1.2
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
python-dotenv==1.2.1
PyYAML==6.0.3
rav==0.2.0
requests==2.32.5
rich==14.2.0
sqlparse==0.5.5
termcolor==3.3.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
django-ninja-jwt[crypto]
